+++ agb/src/display/font.rs
@@ -27,13 +27,19 @@
 //! ```rust
 //! # #![no_std]
 //! # #![no_main]
-//! use agb::display::font::{Layout, AlignmentKind, Font};
+//! use agb::display::font::{Layout, Font, LayoutSettings};
 //!
 //! static FONT: Font = agb::include_font!("examples/font/pixelated.ttf", 8);
 //!
 //! # #[agb::doctest]
 //! # fn test(_: agb::Gba) {
-//! let mut layout = Layout::new("Hello, world!", &FONT, AlignmentKind::Left, 32, 200);
+//! let mut layout = Layout::new(
+//!     "Hello, world!",
+//!     &FONT,
+//!     &LayoutSettings::new()
+//!         .with_max_line_length(200)
+//!         .with_max_group_width(32),
+//! );
 //!
 //! let n = layout.next().unwrap();
 //! assert_eq!(n.text(), "Hello,");
@@ -58,7 +64,7 @@
 //! use alloc::vec::Vec;
 //! use agb::display::{
 //!     Palette16, Rgb15,
-//!     font::{AlignmentKind, Font, Layout, ObjectTextRenderer},
+//!     font::{AlignmentKind, Font, Layout, LayoutSettings, ObjectTextRenderer},
 //!     object::Size,
 //! };
 //!
@@ -74,7 +80,7 @@
 //!
 //! // the actual text rendering
 //!
-//! let layout = Layout::new("Hello, world!", &FONT, AlignmentKind::Left, 16, 200);
+//! let layout = Layout::new("Hello, world!", &FONT, &LayoutSettings::new().with_max_line_length(200));
 //! let text_renderer = ObjectTextRenderer::new(SIMPLE_PALETTE.into(), Size::S16x16);
 //!
 //! for letter_group in layout {
@@ -102,7 +108,7 @@
 //! # #![no_main]
 //! use agb::display::{
 //!     Palette16, Rgb15, Priority,
-//!     font::{AlignmentKind, Font, Layout, RegularBackgroundTextRenderer},
+//!     font::{Font, Layout, LayoutSettings, RegularBackgroundTextRenderer},
 //!     tiled::{RegularBackground, VRAM_MANAGER, RegularBackgroundSize, TileFormat},
 //! };
 //!
@@ -124,7 +130,7 @@
 //!
 //! // the actual text rendering
 //!
-//! let layout = Layout::new("Hello, world!", &FONT, AlignmentKind::Left, 40, 200);
+//! let layout = Layout::new("Hello, world!", &FONT, &LayoutSettings::new().with_max_line_length(200));
 //! let mut text_renderer = RegularBackgroundTextRenderer::new((0, 0));
 //!
 //! for letter_group in layout {
@@ -148,7 +154,7 @@ mod special;
 mod tiled;
 
 pub use align::AlignmentKind;
-pub use layout::{Layout, LetterGroup};
+pub use layout::{Layout, LayoutSettings, LetterGroup};
 pub use object::ObjectTextRenderer;
 pub use tiled::RegularBackgroundTextRenderer;